League and User changes
Most changes were made to the league class. Methods were created to determine  prefixtures for the user team and the AI teams. In the league class saving and loading methods were created to handle the prefixture stacks. 
Changes were made in the User class to save user information and determine whether a user account exists.