Add files via upload
Hey,
I have added the load and save functionalities as well as the initial sign up Screen.
Your username will now be displayed at the top left of the page and the your team name will be displayed as the team name. 
The week is also read in from the text file but it doesn't change yet.